Consolidate expression manipulation functions into optimizer
